+++ confluence/legacy-mobile-macros/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/legacy-mobile-macros
 
+## 2.4.464
+
+### Patch Changes
+
+- Updated dependencies
+
+## 2.4.463
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 2.4.462
 
 ### Patch Changes